LPS-152915 Revert changes of LPS-147918
